requirements.txt
flask
google-cloud-storage==1.27.0
gunicorn

